Pulling the vf dominator out of the pi-nsga-2 code
