+++ pymoo/util/vf_dominator.py
@@ -1,0 +1,69 @@
+import numpy as np
+
+def get_relation(ind_a, ind_b):
+    return Dominator.get_relation(ind_a.F, ind_b.F, ind_a.CV[0], ind_b.CV[0])
+
+
+class VFDominator:
+
+    @staticmethod
+    def get_relation(a, b, cva=None, cvb=None):
+
+        if cva is not None and cvb is not None:
+            if cva < cvb:
+                return 1
+            elif cvb < cva:
+                return -1
+
+        val = 0
+        for i in range(len(a)):
+            if a[i] < b[i]:
+                # indifferent because once better and once worse
+                if val == -1:
+                    return 0
+                val = 1
+            elif b[i] < a[i]:
+                # indifferent because once better and once worse
+                if val == 1:
+                    return 0
+                val = -1
+        return val
+
+    @staticmethod
+    def calc_domination_matrix_loop(F, G):
+        n = F.shape[0]
+        CV = np.sum(G * (G > 0).astype(float), axis=1)
+        M = np.zeros((n, n))
+        for i in range(n):
+            for j in range(i + 1, n):
+                M[i, j] = Dominator.get_relation(F[i, :], F[j, :], CV[i], CV[j])
+                M[j, i] = -M[i, j]
+
+        return M
+
+    @staticmethod
+    def calc_domination_matrix(F, _F=None, epsilon=0.0):
+
+        if _F is None:
+            _F = F
+
+        # look at the obj for dom
+        n = F.shape[0]
+        m = _F.shape[0]
+
+        L = np.repeat(F, m, axis=0)
+        R = np.tile(_F, (n, 1))
+
+        smaller = np.reshape(np.any(L + epsilon < R, axis=1), (n, m))
+        larger = np.reshape(np.any(L > R + epsilon, axis=1), (n, m))
+
+        M = np.logical_and(smaller, np.logical_not(larger)) * 1 \
+            + np.logical_and(larger, np.logical_not(smaller)) * -1
+
+        # if cv equal then look at dom
+        # M = constr + (constr == 0) * dom
+
+        return M
+
+
+
